Update get_dems.sh
the -q flag on wget in get-dems.sh makes it seem like the install script is hanging at this step if you have not great connection speeds to the US servers. --show-progress is way too buggy and verbose with the parallel downloads, but -nv just prints a new line every time a .tif file has completed downloading, showing the user that it's still working.